+++ actstream/apps.py
@@ -1,9 +1,5 @@
-from collections import OrderedDict
-
-import django
-from django.apps import apps
 from django.apps import AppConfig
-from django.conf import settings
+from django.db.models.signals import pre_delete
 
 from actstream import settings as actstream_settings
 from actstream.signals import action
@@ -26,13 +22,5 @@ class ActstreamConfig(AppConfig):
                     action_class, 'data'
                 )
 
-            # dynamically load django_jsonfield_backport to INSTALLED_APPS
-            if django.VERSION < (3, 1) and 'django_jsonfield_backport' not in settings.INSTALLED_APPS:
-                settings.INSTALLED_APPS += ('django_jsonfield_backport', )
-                # reset loaded apps
-                apps.app_configs = OrderedDict()
-                # reset initialization status
-                apps.apps_ready = apps.models_ready = apps.loading = apps.ready = False
-                apps.clear_cache()
-                # re-initialize all apps
-                apps.populate(settings.INSTALLED_APPS)
+        from actstream.follows import delete_orphaned_follows
+        pre_delete.connect(delete_orphaned_follows)