+++ runtests/testapp/apps.py
@@ -1,4 +1,4 @@
-from django.apps import AppConfig
+from django.apps import AppConfig, apps
 
 
 class TestappConfig(AppConfig):
@@ -6,6 +6,8 @@ class TestappConfig(AppConfig):
 
     def ready(self):
         from actstream.registry import register
+        register(apps.get_model('auth', 'group'))
+        register(apps.get_model('sites', 'site'))
         register(self.get_model('player'))
         myuser = self.get_model('myuser')
         if myuser: