+++ runtests/testapp/urls.py
@@ -1,6 +1,8 @@
-from django.urls import re_path
+from django.urls import re_path, path
+from django.conf import settings
 
 from actstream import feeds
+
 
 urlpatterns = [
     re_path(r'custom/(?P<verb>[-\w\s]+)/',